+++ momentum-dao-be/src/main/java/com/dao/momentum/retention/command/domain/aggregate/RetentionSupport.java
@@ -22,7 +22,7 @@ public class RetentionSupport {
     private Long empId;
 
     @Column(name = "round_id", nullable = false)
-    private Long roundId;
+    private Integer roundId;
 
     @Column(name = "rating_score", nullable = false)
     private Integer ratingScore;
